FIX-874: relaxed_iterator/relaxed_range concepts
