+++ include/eve/algo/any_of.hpp
@@ -8,6 +8,7 @@
 #pragma once
 
 #include <eve/algo/array_utils.hpp>
+#include <eve/algo/concepts.hpp>
 #include <eve/algo/for_each_iteration.hpp>
 #include <eve/algo/preprocess_range.hpp>
 #include <eve/algo/traits.hpp>
@@ -48,7 +49,7 @@ namespace eve::algo
       bool res = false;
     };
 
-    template <typename Rng, typename P>
+    template <relaxed_range Rng, typename P>
     EVE_FORCEINLINE bool operator()(Rng&& rng, P p) const
     {
       auto processed = preprocess_range(TraitsSupport::get_traits(), std::forward<Rng>(rng));
