+++ include/eve/algo/concepts/detail.hpp
@@ -1,0 +1,23 @@
+//==================================================================================================
+/*
+  EVE - Expressive Vector Engine
+  Copyright : EVE Contributors & Maintainers
+  SPDX-License-Identifier: MIT
+*/
+//==================================================================================================
+#pragma once
+
+#include <concepts>
+#include <type_traits>
+
+namespace eve::algo::detail
+{
+  template <typename T, template <typename ...> class Templ>
+  struct instance_of_impl : std::false_type {};
+
+  template <typename ...Args, template <typename ...> class Templ>
+  struct instance_of_impl<Templ<Args...>, Templ> : std::true_type {};
+
+  template <typename T, template <typename ...> class Templ>
+  concept instance_of = detail::instance_of_impl<T, Templ>::value;
+}
