+++ include/eve/algo/iterator_helpers.hpp
@@ -7,8 +7,6 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/algo/concepts.hpp>
-
 #include <compare>
 
 namespace eve::algo
