+++ include/eve/algo/zip.hpp
@@ -8,9 +8,12 @@
 #pragma once
 
 #include <eve/algo/as_range.hpp>
+#include <eve/algo/concepts/relaxed.hpp>
 #include <eve/algo/preprocess_range.hpp>
 #include <eve/algo/traits.hpp>
 #include <eve/algo/zip_iterator.hpp>
+
+#include <eve/algo/concepts/detail.hpp>
 
 #include <eve/algo/detail/preprocess_zip_range.hpp>
 
@@ -18,19 +21,12 @@
 
 namespace eve::algo
 {
-  template <typename ZipTraits, typename ...Rngs>
+  template <typename ZipTraits, relaxed_range ...Rngs>
   struct zip_range;
 
   namespace detail
   {
-    // FIX-874: we should do this proper
-    template <typename Rng>
-    concept light_range = requires (Rng& rng) {
-      { rng.begin() };
-      { rng.end() };
-    };
-
-    template <light_range Rng>
+    template <relaxed_range Rng>
     struct rng_ref
     {
       Rng* rng;
@@ -57,7 +53,7 @@ namespace eve::algo
      {
        std::ptrdiff_t res = -1;
        auto process_one = [&]<typename C>(C const& c) mutable {
-         if constexpr (detail::light_range<C>)
+         if constexpr (relaxed_range<C>)
          {
            std::ptrdiff_t cur = c.end() - c.begin();
            if (res == -1) res = cur;
@@ -76,9 +72,9 @@ namespace eve::algo
          [&]<typename C>(C&& c)
          {
            using no_ref = std::remove_reference_t<C>;
-                if constexpr (instance_of<no_ref, detail::rng_ref>) return c;
-           else if constexpr (detail::light_range<no_ref>         ) return detail::rng_ref<no_ref>{&c};
-           else                                                     return as_range(c, unalign(c) + distance);
+                if constexpr (detail::instance_of<no_ref, detail::rng_ref>) return c;
+           else if constexpr (relaxed_range<no_ref>                       ) return detail::rng_ref<no_ref>{&c};
+           else                                                             return as_range(c, unalign(c) + distance);
          }(components)...
        };
      }
@@ -86,9 +82,10 @@ namespace eve::algo
    public:
 
     template <typename ...Components>
+      requires ((relaxed_range<Components> || relaxed_iterator<Components>) && ...)
     EVE_FORCEINLINE auto operator()(Components&& ... components) const
     {
-      if constexpr ((!detail::light_range<Components> && ... ))
+      if constexpr ((relaxed_iterator<Components> && ... ))
       {
         return zip_iterator{components...};
       }
@@ -102,7 +99,7 @@ namespace eve::algo
 
   inline constexpr auto zip = function_with_traits<zip_>;
 
-  template <typename ZipTraits, typename ...Rngs>
+  template <typename ZipTraits, relaxed_range ...Rngs>
   struct zip_range : kumi::tuple<Rngs...>
   {
     ZipTraits zip_tr;
